Fix after review: don't duplicate completed transfers on DB due to duplication of completed transfer on onTransferFinish of ManagerActivityLollipop

+++ app/src/main/java/mega/privacy/android/app/modalbottomsheet/ManageTransferBottomSheetDialogFragment.java
@@ -33,7 +33,7 @@ public class ManageTransferBottomSheetDialogFragment extends BaseBottomSheetDial
 
     private AndroidCompletedTransfer transfer;
     private long handle;
-    private int transferId;
+    private long transferId;
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -188,6 +188,6 @@ public class ManageTransferBottomSheetDialogFragment extends BaseBottomSheetDial
     @Override
     public void onSaveInstanceState(@NonNull Bundle outState) {
         super.onSaveInstanceState(outState);
-        outState.putInt(TRANSFER_ID, transferId);
+        outState.putLong(TRANSFER_ID, transferId);
     }
 }
